requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
numpy==1.24.3
pandas==2.0.3
cplex==22.1.1.0
networkx==3.1
pyscipopt==4.3.0
psutil==5.9.6
py-cpuinfo==9.0.0
tqdm==4.66.1
matplotlib==3.7.2
python-dotenv==1.0.0 